% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{CCT}
\alias{CCT}
\title{Cauchy Combination Test (CCT)}
\usage{
CCT(pvals, weights = NULL)
}
\arguments{
\item{pvals}{p-values, an array of p-values.}

\item{weights}{weights associated with each p-value, SCAMPI used the default weight, 1/len(pvals).}
}
\value{
a single p-value aggregated from pvals.
}
\description{
CCT is used to aggregate multiple p-value developed by Liu and Xie, 202, JASA, 115:529, 393-402, DOI: 10.1080/01621459.2018.1554485
The code used for implementing CCT is adopted from STAAR_v0.9.7 contributed by Xihao Li and Zilin Li at
https://github.com/xihaoli/STAAR/blob/master/R/CCT.R
}
\examples{
set.seed(123)
pval <- runif(20)
CCT(pval)
}
