% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{DgLm}
\alias{DgLm}
\title{Double Generalized Linear Models (DGLM) for Multiple Outcomes}
\usage{
DgLm(y, g_var, z_var = NULL)
}
\arguments{
\item{y}{outcome, a matrix with multiple columns, representing multiple traits}

\item{g_var}{genotype, column matrix G with one column, representing one genotype variable}

\item{z_var}{optional confounder, matrix Z with multiple columns, representing multiple confounders}
}
\value{
standardized traits of the original trait matrix y, a matrix with multiple columns
}
\description{
Apply Double Generalized Linear Models (DGLM) to standardize each
column in matrix y. The mean sub-model will be standardized by one genotype variable g_var and multiple confounders z_var;
The variance sub-model will be standardized by confounders z_var. The confounders z_var is optional.
}
\examples{
DgLm(
  y = matrix(rnorm(500), ncol = 5),
  z_var = matrix(rnorm(400), ncol = 4),
  g_var = rbinom(100, 2, 0.25)
)
}
