% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{levene_test}
\alias{levene_test}
\title{Univariate Levene's Test}
\usage{
levene_test(y, g_var)
}
\arguments{
\item{y}{outcome, column matrix, representing one single trait}

\item{g_var}{genotype, column matrix G with one column, representing one genotype variable}
}
\value{
one single p-value
}
\description{
The univariate Levene's test can be applied to a single trait. It is implemented
based on the formula in Pare and Cook, 2010, PLoS Genet 6(6): e1000981. doi:10.1371/journal.pgen.1000981
}
\examples{
set.seed(123)
N <- 1000
G <- matrix(rbinom(N, 2, 0.25), ncol = 1)
W <- matrix(rnorm(N), ncol = 1)
Y <- 0.2 + 0.1 * G + 0.3 * W + 2 * G * W + rnorm(N)
levene_test(y = Y, g_var = G)
}
