% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{multivariate_levene_test}
\alias{multivariate_levene_test}
\title{Multivariate Levene's Test}
\usage{
multivariate_levene_test(y, g_var, z_var = NULL)
}
\arguments{
\item{y}{outcome, a matrix with multiple columns, representing multiple traits}

\item{g_var}{genotype, column matrix G with one column, representing one genotype variable}

\item{z_var}{optional confounder, matrix Z with multiple columns, representing multiple confounders}
}
\value{
one aggregated p-value for Multivariate Levene's Test
}
\description{
Multivariate Levene's Test is extended in our paper from the univariate Levene's Test to
accomodate multiple traits.
}
\examples{
set.seed(123)
multivariate_levene_test(
  y = matrix(rnorm(500), ncol = 5),
  z_var = matrix(rnorm(400), ncol = 4),
  g_var = matrix(rbinom(100, 2, 0.25), ncol = 1)
)
}
