% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{PC_scampi_test}
\alias{PC_scampi_test}
\title{PC SCAMPI}
\usage{
PC_scampi_test(y, g_var, z_var = NULL)
}
\arguments{
\item{y}{outcome, a matrix with multiple columns, representing multiple traits}

\item{g_var}{genotype, column matrix G with one column, representing one genotype variable}

\item{z_var}{optional confounder, matrix Z with multiple columns, representing multiple confounders}
}
\value{
PC SCAMPI p-value
}
\description{
This is a variation of the SCAMPI method. PC SCAMPI treats the PC of traits as
the outcomes.
}
\examples{
set.seed(123)
PC_scampi_test(
  y = matrix(rnorm(500), ncol = 5),
  z_var = matrix(rnorm(400), ncol = 4),
  g_var = matrix(rbinom(100, 2, 0.25), ncol = 1)
)
}
