% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAMPI.R
\name{SCAMPI}
\alias{SCAMPI}
\title{Scalable Cauchy Aggregate test using Multiple Phenotypes to test Interactions (SCAMPI)}
\usage{
SCAMPI(y, g_var, z_var = NULL)
}
\arguments{
\item{y}{outcome, a matrix with multiple columns, representing multiple traits}

\item{g_var}{genotype, column matrix G with one column, representing one genotype variable}

\item{z_var}{optional confounder, matrix Z with multiple columns, representing multiple confounders}
}
\value{
SCAMPI p-value
}
\description{
SCAMPI method is applied for detecting the G x E or G x G interaction effects by
utilizing both variance and covariance structure of multiple traits.
}
\examples{
SCAMPI(
  y = matrix(rnorm(500), ncol = 5),
  z_var = matrix(rnorm(400), ncol = 4),
  g_var = rbinom(100, 2, 0.25)
)
}
